tantra: support long long hex printing

+++ src/fb.c
@@ -54,7 +54,7 @@ static inline void fb_scroll()
 }
 
 static const char *hexs = "0123456789abcdef";
-void fb_puthex(unsigned long h)
+void fb_puthex(unsigned long long h)
 {
     if (h <= 0)
     {
